# public/swagger.json
{
	"openapi": "3.0.0",
	"components": {
		"examples": {},
		"headers": {},
		"parameters": {},
		"requestBodies": {},
		"responses": {},
		"schemas": {
			"Gender": {
				"enum": [
					"male",
					"female",
					"other",
					"unknown"
				],
				"type": "string"
			},
			"User": {
				"properties": {
					"id": {
						"type": "string"
					},
					"firstName": {
						"type": "string"
					},
					"lastName": {
						"type": "string"
					},
					"email": {
						"type": "string"
					},
					"password": {
						"type": "string"
					},
					"gender": {
						"$ref": "#/components/schemas/Gender"
					},
					"isDeleted": {
						"type": "boolean",
						"default": false
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					},
					"updatedAt": {
						"type": "string",
						"format": "date-time"
					},
					"sesssions": {
						"items": {
							"$ref": "#/components/schemas/Session"
						},
						"type": "array"
					},
					"userRoles": {
						"items": {
							"$ref": "#/components/schemas/UserRole"
						},
						"type": "array"
					}
				},
				"required": [
					"id",
					"firstName",
					"lastName",
					"email",
					"password",
					"createdAt",
					"updatedAt",
					"sesssions",
					"userRoles"
				],
				"type": "object",
				"additionalProperties": false
			},
			"SessionStatus": {
				"enum": [
					"active",
					"expired",
					"terminated"
				],
				"type": "string"
			},
			"Session": {
				"properties": {
					"id": {
						"type": "string"
					},
					"user": {
						"$ref": "#/components/schemas/User"
					},
					"accessToken": {
						"type": "string"
					},
					"refreshToken": {
						"type": "string"
					},
					"expireAt": {
						"type": "string",
						"format": "date-time"
					},
					"isDeleted": {
						"type": "boolean",
						"default": false
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					},
					"updatedAt": {
						"type": "string",
						"format": "date-time"
					},
					"lastAccessed": {
						"type": "string",
						"format": "date-time"
					},
					"status": {
						"$ref": "#/components/schemas/SessionStatus"
					}
				},
				"required": [
					"id",
					"user",
					"accessToken",
					"expireAt",
					"createdAt",
					"updatedAt"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Roles": {
				"enum": [
					"user",
					"admin",
					"super_admin"
				],
				"type": "string"
			},
			"UserRole": {
				"properties": {
					"userId": {
						"type": "string"
					},
					"roleId": {
						"type": "string"
					},
					"user": {
						"$ref": "#/components/schemas/User"
					},
					"role": {
						"$ref": "#/components/schemas/Role"
					}
				},
				"required": [
					"userId",
					"roleId",
					"user",
					"role"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Role": {
				"properties": {
					"id": {
						"type": "string"
					},
					"name": {
						"$ref": "#/components/schemas/Roles"
					},
					"description": {
						"type": "string"
					},
					"userRoles": {
						"items": {
							"$ref": "#/components/schemas/UserRole"
						},
						"type": "array"
					}
				},
				"required": [
					"id",
					"userRoles"
				],
				"type": "object",
				"additionalProperties": false
			},
			"TokenResponseDTO": {
				"properties": {
					"accessToken": {
						"type": "string"
					},
					"refreshToken": {
						"type": "string"
					},
					"expireAt": {
						"type": "string",
						"format": "date-time"
					}
				},
				"required": [
					"accessToken",
					"expireAt"
				],
				"type": "object",
				"additionalProperties": false
			},
			"SuccessResponse_TokenResponseDTO_": {
				"properties": {
					"code": {
						"anyOf": [
							{
								"type": "string"
							},
							{
								"type": "string",
								"enum": [
									""
								]
							}
						]
					},
					"message": {
						"type": "string"
					},
					"result": {
						"$ref": "#/components/schemas/TokenResponseDTO"
					}
				},
				"required": [
					"code",
					"message",
					"result"
				],
				"type": "object",
				"additionalProperties": false
			},
			"IRegisterUser": {
				"properties": {
					"firstName": {
						"type": "string"
					},
					"lastName": {
						"type": "string"
					},
					"email": {
						"type": "string"
					},
					"gender": {
						"$ref": "#/components/schemas/Gender"
					},
					"password": {
						"type": "string"
					},
					"roles": {
						"items": {
							"type": "string"
						},
						"type": "array"
					}
				},
				"required": [
					"firstName",
					"lastName",
					"email",
					"gender",
					"password",
					"roles"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ILoginUser": {
				"properties": {
					"email": {
						"type": "string"
					},
					"password": {
						"type": "string"
					}
				},
				"required": [
					"email",
					"password"
				],
				"type": "object",
				"additionalProperties": false
			},
			"SuccessResponse_null_": {
				"properties": {
					"code": {
						"anyOf": [
							{
								"type": "string"
							},
							{
								"type": "string",
								"enum": [
									""
								]
							}
						]
					},
					"message": {
						"type": "string"
					},
					"result": {
						"type": "number",
						"enum": [
							null
						],
						"nullable": true
					}
				},
				"required": [
					"code",
					"message",
					"result"
				],
				"type": "object",
				"additionalProperties": false
			},
			"UserProfileDTO": {
				"properties": {
					"firstName": {
						"type": "string"
					},
					"lastName": {
						"type": "string"
					},
					"email": {
						"type": "string"
					},
					"gender": {
						"$ref": "#/components/schemas/Gender"
					}
				},
				"required": [
					"firstName",
					"lastName",
					"email"
				],
				"type": "object",
				"additionalProperties": false
			},
			"SuccessResponse_UserProfileDTO-Array_": {
				"properties": {
					"code": {
						"anyOf": [
							{
								"type": "string"
							},
							{
								"type": "string",
								"enum": [
									""
								]
							}
						]
					},
					"message": {
						"type": "string"
					},
					"result": {
						"items": {
							"$ref": "#/components/schemas/UserProfileDTO"
						},
						"type": "array"
					}
				},
				"required": [
					"code",
					"message",
					"result"
				],
				"type": "object",
				"additionalProperties": false
			},
			"SuccessResponse_UserProfileDTO_": {
				"properties": {
					"code": {
						"anyOf": [
							{
								"type": "string"
							},
							{
								"type": "string",
								"enum": [
									""
								]
							}
						]
					},
					"message": {
						"type": "string"
					},
					"result": {
						"$ref": "#/components/schemas/UserProfileDTO"
					}
				},
				"required": [
					"code",
					"message",
					"result"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Partial_IUpdateUser_": {
				"properties": {
					"firstName": {
						"type": "string"
					},
					"lastName": {
						"type": "string"
					},
					"gender": {
						"$ref": "#/components/schemas/Gender"
					}
				},
				"type": "object",
				"description": "Make all properties in T optional"
			}
		},
		"securitySchemes": {
			"api_key": {
				"type": "apiKey",
				"name": "x-access-token",
				"in": "header",
				"description": "API Key Authentication"
			},
			"jwt": {
				"type": "apiKey",
				"name": "x-access-token",
				"in": "header"
			}
		}
	},
	"info": {
		"title": "authenticationwithkeycloak",
		"version": "1.0.0",
		"license": {
			"name": "MIT"
		},
		"contact": {}
	},
	"paths": {
		"/api/v1/health": {
			"get": {
				"operationId": "CheckHealth",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/User"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"tags": [
					"Health"
				],
				"security": [],
				"parameters": []
			}
		},
		"/api/v1/auth/register": {
			"post": {
				"operationId": "Register",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/SuccessResponse_TokenResponseDTO_"
								}
							}
						}
					}
				},
				"tags": [
					"Authentication"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/IRegisterUser"
							}
						}
					}
				}
			}
		},
		"/api/v1/auth/login": {
			"post": {
				"operationId": "Login",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/SuccessResponse_TokenResponseDTO_"
								}
							}
						}
					}
				},
				"tags": [
					"Authentication"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/ILoginUser"
							}
						}
					}
				}
			}
		},
		"/api/v1/auth/logout": {
			"get": {
				"operationId": "Logout",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/SuccessResponse_null_"
								}
							}
						}
					}
				},
				"tags": [
					"Authentication"
				],
				"security": [],
				"parameters": [
					{
						"in": "header",
						"name": "x-access-token",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/api/v1/users": {
			"get": {
				"operationId": "GetAllUsers",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/SuccessResponse_UserProfileDTO-Array_"
								}
							}
						}
					}
				},
				"tags": [
					"Admin"
				],
				"security": [
					{
						"jwt": [
							"read:profile",
							"write:profile"
						]
					}
				],
				"parameters": []
			}
		},
		"/api/v1/users/{userId}": {
			"get": {
				"operationId": "GetUser",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/SuccessResponse_UserProfileDTO_"
								}
							}
						}
					}
				},
				"tags": [
					"Admin"
				],
				"security": [
					{
						"jwt": [
							"read:profile",
							"write:profile"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "userId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/api/v1/users/{userId}/delete": {
			"delete": {
				"operationId": "DeleteUser",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/SuccessResponse_null_"
								}
							}
						}
					}
				},
				"tags": [
					"Admin"
				],
				"security": [
					{
						"jwt": [
							"read:profile",
							"write:profile"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "userId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/api/v1/profile": {
			"get": {
				"operationId": "GetProfile",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/SuccessResponse_UserProfileDTO_"
								}
							}
						}
					}
				},
				"tags": [
					"User"
				],
				"security": [
					{
						"jwt": [
							"read:profile"
						]
					}
				],
				"parameters": []
			}
		},
		"/api/v1/update": {
			"put": {
				"operationId": "UpdateProfile",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/SuccessResponse_UserProfileDTO_"
								}
							}
						}
					}
				},
				"tags": [
					"User"
				],
				"security": [
					{
						"jwt": [
							"read:profile",
							"write:profile"
						]
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/Partial_IUpdateUser_"
							}
						}
					}
				}
			}
		}
	},
	"servers": [
		{
			"url": "/"
		}
	]
}